test(eve): isolate compaction model regressions
