+++ e2e/fixtures/agent-compaction-regressions-gpt-5-6/agent/instructions.md
