+++ e2e/fixtures/agent-compaction-regressions-gpt-5-6/agent/tools/inspect-repository.ts
@@ -1,0 +1,1 @@
+export { default } from "@eve/e2e-compaction-regression-shared/tools/inspect-repository";
