+++ e2e/fixtures/agent-compaction-regressions-gpt-5-6/agent/tools/perform-source-analysis.ts
@@ -1,0 +1,1 @@
+export { default } from "@eve/e2e-compaction-regression-shared/tools/perform-source-analysis";
