+++ e2e/fixtures/agent-compaction-regressions-gpt-5-6/evals/redundant-tool-calls/gpt-5.6.eval.ts
@@ -1,6 +1,5 @@
 import { defineEval } from "eve/evals";
-
-import { testRedundantToolCalls } from "../shared";
+import { testRedundantToolCalls } from "@eve/e2e-compaction-regression-shared/evals";
 
 export default defineEval({
   description: "GPT-5.6 does not repeat an identical successful call after compaction.",
