+++ e2e/fixtures/agent-compaction-regressions-gpt-5-6/evals/stale-todo-work/gpt-5.6.eval.ts
@@ -1,6 +1,5 @@
 import { defineEval } from "eve/evals";
-
-import { testStaleTodoWork } from "../shared";
+import { testStaleTodoWork } from "@eve/e2e-compaction-regression-shared/evals";
 
 export default defineEval({
   description: "GPT-5.6 does not repeat completed work because a todo stayed pending.",
