+++ e2e/fixtures/agent-compaction-regressions-gpt-5-6/tsconfig.json
