+++ e2e/fixtures/agent-compaction-regressions-opus-4-8/evals/evals.config.ts
@@ -1,0 +1,6 @@
+import { defineEvalConfig } from "eve/evals";
+
+export default defineEvalConfig({
+  maxConcurrency: 1,
+  timeoutMs: 180_000,
+});
