+++ e2e/fixtures/agent-compaction-regressions-opus-4-8/evals/redundant-tool-calls/opus-4.8.eval.ts
@@ -1,6 +1,5 @@
+import { testRedundantToolCalls } from "@eve/e2e-compaction-regression-shared/evals";
 import { defineEval } from "eve/evals";
-
-import { testRedundantToolCalls } from "../shared";
 
 export default defineEval({
   description: "Claude Opus 4.8 does not repeat an identical successful call after compaction.",
