+++ e2e/fixtures/agent-compaction-regressions-opus-4-8/evals/stale-todo-work/opus-4.8.eval.ts
@@ -1,6 +1,5 @@
+import { testStaleTodoWork } from "@eve/e2e-compaction-regression-shared/evals";
 import { defineEval } from "eve/evals";
-
-import { testStaleTodoWork } from "../shared";
 
 export default defineEval({
   description: "Claude Opus 4.8 does not repeat completed work because a todo stayed pending.",
