+++ e2e/fixtures/agent-compaction-regressions-sonnet-5/evals/redundant-tool-calls/sonnet-5.eval.ts
@@ -1,6 +1,5 @@
+import { testRedundantToolCalls } from "@eve/e2e-compaction-regression-shared/evals";
 import { defineEval } from "eve/evals";
-
-import { testRedundantToolCalls } from "../shared";
 
 export default defineEval({
   description: "Claude Sonnet 5 does not repeat an identical successful call after compaction.",
