+++ e2e/fixtures/agent-compaction-regressions-sonnet-5/evals/stale-todo-work/sonnet-5.eval.ts
@@ -1,6 +1,5 @@
+import { testStaleTodoWork } from "@eve/e2e-compaction-regression-shared/evals";
 import { defineEval } from "eve/evals";
-
-import { testStaleTodoWork } from "../shared";
 
 export default defineEval({
   description: "Claude Sonnet 5 does not repeat completed work because a todo stayed pending.",
