+++ e2e/fixtures/agent-compaction-regressions-sonnet-5/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "agent-compaction-regressions",
+  "name": "agent-compaction-regressions-sonnet-5",
   "version": "0.0.0",
   "private": true,
   "type": "module",
@@ -11,10 +11,10 @@
     "test:e2e": "eve eval --strict"
   },
   "dependencies": {
+    "@eve/e2e-compaction-regression-shared": "workspace:*",
     "@opentelemetry/core": "2.6.1",
     "@opentelemetry/sdk-trace-base": "2.6.1",
     "@vercel/otel": "2.1.2",
-    "ai": "catalog:",
     "eve": "workspace:*",
     "zod": "catalog:"
   },
